[fix] Fix caret position, use logic of ep_autocomp
Use same logic of ep_autocomp to calculate position of caret indicator,
as that plugin has several tests and has a more modular code for that.

Also fix some scenarios where logic was incorrect:
- lines with line attributes were placing caret indicator one char to
the right (was not ignoring the '*' on the beginning of line);
- on divs with right padding, caret was not being placed on next line
when caret was at the end of the line;

There are still some scenarios to fix, this is not a final commit for
these fixes.