[fix] Avoid `splitText` error when loading the pad (#16)
Etherpad loads the caret at position [0,0], even when the first line of
pad has the line marker ('*'). This causes the plugin to calculate the
caret column as -1, which will cause a `splitText` error later. To avoid
that, we fix the calculated column and manually set it to 0.

This might fix [bug #1514](https://trello.com/c/9ug7o5Zg/1514), but most likely it
only fixes *some* cases of the error.